+++ packages/clerk-js/src/ui/common/constants.ts
@@ -1,50 +1,47 @@
 import type { OAuthProvider, Web3Provider } from '@clerk/types';
 
-const FirstFactorConfigs = Object.freeze({
+export const FirstFactorConfigs = Object.freeze({
   email_address: {
     label: 'Email address',
+    icon: 'clerk-email',
     fieldType: 'email',
   },
   phone_number: {
     label: 'Phone number',
+    icon: 'clerk-phone',
     fieldType: 'tel',
   },
   username: {
     label: 'Username',
+    icon: 'clerk-person',
     fieldType: 'text',
   },
   email_address_phone_number: {
     label: 'Email or phone',
+    icon: 'clerk-person',
     fieldType: 'text',
   },
   email_address_username: {
     label: 'Email or username',
+    icon: 'clerk-person',
     fieldType: 'text',
   },
   phone_number_username: {
     label: 'Phone number or username',
+    icon: 'clerk-person',
     fieldType: 'text',
   },
   email_address_phone_number_username: {
     label: 'Email, phone, or username',
+    icon: 'clerk-person',
     fieldType: 'text',
   },
-  default: {
-    label: '',
-    fieldType: 'text',
-  },
-} as Record<string, { label: string; fieldType: string }>);
-
-export const getIdentifierControlDisplayValues = (attributes: string[]) => {
-  const indexKey = attributes.length == 0 ? null : [...attributes].sort().join('_');
-  return FirstFactorConfigs[indexKey || 'default'];
-};
+} as Record<string, { label: string; icon: string; fieldType: string }>);
 
 export const PREFERRED_SIGN_IN_STRATEGIES = Object.freeze({
   Password: 'password',
   OTP: 'otp',
 });
-
 interface OAuthProviderData {
   id: string;
   name: string;
@@ -103,9 +100,15 @@ export const OAUTH_PROVIDERS: OAuthProviders = Object.freeze({
     id: 'bitbucket',
     name: 'Bitbucket',
   },
+  microsoft: {
+    id: 'microsoft',
+    name: 'Microsoft',
+  },
 });
 
-export function getOAuthProviderData(name: OAuthProvider): OAuthProviderData | undefined | null {
+export function getOAuthProviderData(
+  name: OAuthProvider,
+): OAuthProviderData | undefined | null {
   return OAUTH_PROVIDERS[name];
 }
 
@@ -125,7 +128,9 @@ export const WEB3_PROVIDERS: Web3Providers = Object.freeze({
   },
 });
 
-export function getWeb3ProviderData(name: Web3Provider): Web3ProviderData | undefined | null {
+export function getWeb3ProviderData(
+  name: Web3Provider,
+): Web3ProviderData | undefined | null {
   return WEB3_PROVIDERS[name];
 }