+++ packages/clerk-js/src/ui/signIn/SignInStart.test.tsx
@@ -1,7 +1,14 @@
-import { mocked, render, renderJSON, screen, userEvent, waitFor } from '@clerk/shared/testUtils';
+import {
+  mocked,
+  render,
+  renderJSON,
+  screen,
+  userEvent,
+  waitFor,
+} from '@clerk/shared/testUtils';
 import { titleize } from '@clerk/shared/utils/string';
-import { EnvironmentResource, SignInResource, UserSettingsJSON, UserSettingsResource } from '@clerk/types';
-import { ClerkAPIResponseError, UserSettings } from 'core/resources/internal';
+import { EnvironmentResource, SignInResource } from '@clerk/types';
+import { ClerkAPIResponseError } from 'core/resources/Error';
 import React from 'react';
 import { useCoreSignIn } from 'ui/contexts';
 
@@ -11,7 +18,6 @@ const mockNavigate = jest.fn();
 const mockCreateRequest = jest.fn();
 const mockAuthenticateWithRedirect = jest.fn();
 const mockNavigateAfterSignIn = jest.fn();
-let mockUserSettings: UserSettingsResource;
 
 jest.mock('ui/router/RouteContext');
 
@@ -36,9 +42,15 @@ jest.mock('ui/contexts', () => {
             preferredSignInStrategy: 'otp',
             afterSignInUrl: 'http://test.host',
           },
-          userSettings: mockUserSettings,
           authConfig: {
+            password: 'required',
             singleSessionMode: false,
+            identificationStrategies: [
+              'email_address',
+              'oauth_google',
+              'oauth_facebook',
+            ],
+            firstFactors: ['email_address', 'oauth_google', 'oauth_facebook'],
           },
         } as any as EnvironmentResource),
     ),
@@ -69,44 +81,9 @@ jest.mock('ui/hooks', () => ({
   },
 }));
 
-fdescribe('<SignInStart/>', () => {
+describe('<SignInStart/>', () => {
   beforeEach(() => {
     jest.clearAllMocks();
-  });
-
-  beforeEach(() => {
-    mockUserSettings = new UserSettings({
-      attributes: {
-        email_address: {
-          enabled: true,
-          required: false,
-          used_for_first_factor: true,
-          first_factors: ['email_link'],
-          used_for_second_factor: false,
-          second_factors: [],
-          verifications: ['email_link'],
-          verify_at_sign_up: false,
-        },
-        password: {
-          enabled: true,
-          used_for_first_factor: false,
-        },
-      },
-      social: {
-        oauth_google: {
-          enabled: true,
-          required: false,
-          authenticatable: false,
-          strategy: 'oauth_google',
-        },
-        oauth_facebook: {
-          enabled: true,
-          required: false,
-          authenticatable: false,
-          strategy: 'oauth_facebook',
-        },
-      },
-    } as unknown as UserSettingsJSON);
   });
 
   describe('when user is not signed in', () => {
@@ -117,7 +94,7 @@ fdescribe('<SignInStart/>', () => {
       });
     });
 
-    it('renders the sign in start screen', () => {
+    it('renders the sign in start screen', async () => {
       const tree = renderJSON(<SignInStart />);
       expect(tree).toMatchSnapshot();
     });
@@ -126,7 +103,7 @@ fdescribe('<SignInStart/>', () => {
       render(<SignInStart />);
 
       const inputField = screen.getByLabelText('Email address');
-      userEvent.type(inputField, 'boss@clerk.dev');
+      await userEvent.type(inputField, 'boss@clerk.dev');
 
       const signEmailButton = screen.getByRole('button', { name: /Continue/i });
       userEvent.click(signEmailButton);
@@ -147,15 +124,17 @@ fdescribe('<SignInStart/>', () => {
         status: 'complete',
       });
 
-      const instantPasswordField = container.querySelector('input#password') as HTMLInputElement;
+      const instantPasswordField = container.querySelector(
+        'input#password',
+      ) as HTMLInputElement;
 
       expect(instantPasswordField).toBeDefined();
 
       const inputField = screen.getByLabelText('Email address');
-      userEvent.type(inputField, 'boss@clerk.dev');
+      await userEvent.type(inputField, 'boss@clerk.dev');
 
       // simulate password being filled by a pwd manager
-      userEvent.type(instantPasswordField, '123456');
+      await userEvent.type(instantPasswordField, '123456');
 
       const signEmailButton = screen.getByRole('button', { name: /Continue/i });
       userEvent.click(signEmailButton);
@@ -190,15 +169,17 @@ fdescribe('<SignInStart/>', () => {
           status: 'needs_first_factor',
         }));
 
-      const instantPasswordField = container.querySelector('input#password') as HTMLInputElement;
+      const instantPasswordField = container.querySelector(
+        'input#password',
+      ) as HTMLInputElement;
 
       expect(instantPasswordField).toBeDefined();
 
       const inputField = screen.getByLabelText('Email address');
-      await userEvent.type(inputField, 'boss@clerk.dev');
+      userEvent.type(inputField, 'boss@clerk.dev');
 
       // simulate password being filled by a pwd manager
-      await userEvent.type(instantPasswordField, 'wrong pass');
+      userEvent.type(instantPasswordField, 'wrong pass');
 
       const signEmailButton = screen.getByRole('button', { name: /Continue/i });
       userEvent.click(signEmailButton);
@@ -228,7 +209,8 @@ fdescribe('<SignInStart/>', () => {
             data: [
               {
                 code: 'form_password_incorrect',
-                message: 'Password is incorrect. Try again, or use another method.',
+                message:
+                  'Password is incorrect. Try again, or use another method.',
                 meta: {
                   param_name: 'password',
                 },
@@ -241,7 +223,9 @@ fdescribe('<SignInStart/>', () => {
           status: 'needs_first_factor',
         }));
 
-      const instantPasswordField = container.querySelector('input#password') as HTMLInputElement;
+      const instantPasswordField = container.querySelector(
+        'input#password',
+      ) as HTMLInputElement;
 
       expect(instantPasswordField).toBeDefined();
 
@@ -332,16 +316,24 @@ fdescribe('<SignInStart/>', () => {
           identifier: 'boss@clerk.dev',
         });
         expect(mockNavigate).not.toHaveBeenCalled();
-        expect(mockSetSession).toHaveBeenNthCalledWith(1, 'deadbeef', mockNavigateAfterSignIn);
+        expect(mockSetSession).toHaveBeenNthCalledWith(
+          1,
+          'deadbeef',
+          mockNavigateAfterSignIn,
+        );
       });
     });
   });
 
   describe('when the instance is invitation only', () => {
     it('renders the external account verification error if available', async () => {
-      const errorMsg = 'You cannot sign up with sokratis.vidros@gmail.com since this is an invitation-only application';
+      const errorMsg =
+        'You cannot sign up with sokratis.vidros@gmail.com since this is an invitation-only application';
 
-      mocked(useCoreSignIn as jest.Mock<SignInResource>, true).mockImplementationOnce(
+      mocked(
+        useCoreSignIn as jest.Mock<SignInResource>,
+        true,
+      ).mockImplementationOnce(
         () =>
           ({
             create: mockCreateRequest,
@@ -365,7 +357,10 @@ fdescribe('<SignInStart/>', () => {
     it('renders the external account verification error if available', async () => {
       const errorMsg = 'You did not grant access to your Google account';
 
-      mocked(useCoreSignIn as jest.Mock<SignInResource>, true).mockImplementationOnce(
+      mocked(
+        useCoreSignIn as jest.Mock<SignInResource>,
+        true,
+      ).mockImplementationOnce(
         () =>
           ({
             create: mockCreateRequest,