+++ packages/clerk-js/src/ui/signIn/SignInStart.tsx
@@ -1,14 +1,19 @@
 import { Control } from '@clerk/shared/components/control';
 import { Form } from '@clerk/shared/components/form';
-import { Input } from '@clerk/shared/components/input';
+import { Input, InputType } from '@clerk/shared/components/input';
 import { PhoneInput } from '@clerk/shared/components/phoneInput';
-import { ClerkAPIError, SignInParams } from '@clerk/types';
+import {
+  ClerkAPIError,
+  OAuthStrategy,
+  SignInParams,
+  Web3Strategy,
+} from '@clerk/types';
 import cn from 'classnames';
 import React from 'react';
 import {
   buildRequest,
   FieldState,
-  getIdentifierControlDisplayValues,
+  FirstFactorConfigs,
   handleError,
   LoadingScreen,
   PoweredByClerk,
@@ -32,7 +37,7 @@ import { SignUpLink } from './SignUpLink';
 import { OAuth, Web3 } from './strategies';
 
 export function _SignInStart(): JSX.Element {
-  const { userSettings } = useEnvironment();
+  const environment = useEnvironment();
   const { setSession } = useCoreClerk();
   const signIn = useCoreSignIn();
   const { navigate } = useNavigate();
@@ -45,13 +50,7 @@ export function _SignInStart(): JSX.Element {
   const [error, setError] = React.useState<string | undefined>();
   const [isLoading, setIsLoading] = React.useState(false);
 
-  const standardFormAttributes = userSettings.enabledFirstFactorIdentifiers;
-  const web3FirstFactors = userSettings.web3FirstFactors;
-  const socialProviderStrategies = userSettings.socialProviderStrategies;
-
-  const identifierInputDisplayValues = getIdentifierControlDisplayValues(
-    standardFormAttributes,
-  );
+  const { authConfig } = environment;
 
   React.useEffect(() => {
     if (!organizationTicket) {
@@ -86,6 +85,37 @@ export function _SignInStart(): JSX.Element {
       });
   }, []);
 
+  const firstPartyOptions = authConfig.identificationStrategies.filter(
+    strategy => !strategy.includes('oauth') && !strategy.includes('web3'),
+  );
+
+  const firstPartyKey =
+    firstPartyOptions.length == 0
+      ? null
+      : [...firstPartyOptions].sort().join('_');
+
+  const firstPartyLabel =
+    firstPartyKey && FirstFactorConfigs[firstPartyKey]
+      ? FirstFactorConfigs[firstPartyKey].label
+      : '';
+
+  const fieldType: InputType = (
+    firstPartyKey && FirstFactorConfigs[firstPartyKey]
+      ? FirstFactorConfigs[firstPartyKey].fieldType
+      : 'text'
+  ) as InputType;
+
+  const firstFactors = authConfig.firstFactors;
+  const web3Options = firstFactors
+    .filter(fac => fac.startsWith('web3'))
+    .sort() as Web3Strategy[];
+  const oauthOptions = firstFactors
+    .filter(fac => fac.startsWith('oauth'))
+    .sort() as OAuthStrategy[];
+
+  const passwordBasedInstance = authConfig.password === 'required';
+
+  // TODO: Clean up the following code end
   React.useEffect(() => {
     async function handleOauthError() {
       const error = signIn?.firstFactorVerification?.error;
@@ -94,8 +124,11 @@ export function _SignInStart(): JSX.Element {
         error?.code === ERROR_CODES.OAUTH_ACCESS_DENIED
       ) {
         setError(error.longMessage);
-        // TODO: This is a workaround in order to reset the sign in attempt
-        // so that the oauth error does not persist on full page reloads.
+
+        // TODO: This is a hack to reset the sign in attempt so that the oauth error
+        // does not persist on full page reloads.
+        //
+        // We will revise this strategy as part of the Clerk DX epic.
         void (await signIn.create({}));
       }
     }
@@ -166,9 +199,6 @@ export function _SignInStart(): JSX.Element {
     return signInWithFields(identifier, instantPassword);
   };
 
-  const hasSocialOrWeb3Buttons =
-    !!socialProviderStrategies.length || !!web3FirstFactors.length;
-
   if (isLoading) {
     return <LoadingScreen />;
   }
@@ -177,32 +207,26 @@ export function _SignInStart(): JSX.Element {
     <>
       <Header error={error} />
       <Body>
-        <OAuth
-          oauthOptions={socialProviderStrategies}
-          setError={setError}
-          error={error}
-        />
-        <Web3
-          web3Options={web3FirstFactors}
-          setError={setError}
-          error={error}
-        />
+        <OAuth oauthOptions={oauthOptions} setError={setError} error={error} />
+        <Web3 web3Options={web3Options} setError={setError} error={error} />
 
-        {standardFormAttributes.length > 0 && (
+        {firstPartyOptions.length > 0 && (
           <>
-            {hasSocialOrWeb3Buttons && <Separator />}
+            {(oauthOptions.length > 0 || web3Options.length > 0) && (
+              <Separator />
+            )}
             <Form
               handleSubmit={handleFirstPartySubmit}
               submitButtonClassName='cl-sign-in-button'
               submitButtonLabel='Continue'
             >
               <Control
-                label={identifierInputDisplayValues.label}
+                label={firstPartyLabel}
                 labelClassName='cl-label'
                 error={identifier.error}
                 htmlFor='text-field-identifier'
               >
-                {identifierInputDisplayValues.fieldType === phoneFieldType ? (
+                {fieldType === phoneFieldType ? (
                   <PhoneInput
                     id='text-field-identifier'
                     name='text-field-identifier'
@@ -212,7 +236,7 @@ export function _SignInStart(): JSX.Element {
                   <Input
                     id='text-field-identifier'
                     name='text-field-identifier'
-                    type={identifierInputDisplayValues.fieldType as any}
+                    type={fieldType}
                     handleChange={el => identifier.setValue(el.value || '')}
                     value={identifier.value}
                     autoFocus
@@ -222,24 +246,30 @@ export function _SignInStart(): JSX.Element {
                 )}
               </Control>
 
-              <Control
-                key='password'
-                htmlFor='password'
-                label='Password'
-                error={instantPassword.error}
-                className={cn({
-                  'cl-hidden': !instantPassword.value,
-                })}
-              >
-                <Input
-                  id='password'
-                  type='password'
-                  name='password'
-                  value={instantPassword.value}
-                  handleChange={el => instantPassword.setValue(el.value || '')}
-                  tabIndex={-1}
-                />
-              </Control>
+              <>
+                {passwordBasedInstance && (
+                  <Control
+                    key='password'
+                    htmlFor='password'
+                    label='Password'
+                    error={instantPassword.error}
+                    className={cn({
+                      'cl-hidden': !instantPassword.value,
+                    })}
+                  >
+                    <Input
+                      id='password'
+                      type='password'
+                      name='password'
+                      value={instantPassword.value}
+                      handleChange={el =>
+                        instantPassword.setValue(el.value || '')
+                      }
+                      tabIndex={-1}
+                    />
+                  </Control>
+                )}
+              </>
             </Form>
           </>
         )}