+++ packages/clerk-js/src/ui/signUp/SignUp.test.tsx
@@ -1,6 +1,7 @@
 import { render } from '@clerk/shared/testUtils';
 import { EnvironmentResource } from '@clerk/types';
-import { AuthConfig, Session } from 'core/resources/internal';
+import { AuthConfig } from 'core/resources/AuthConfig';
+import { Session } from 'core/resources/Session';
 import React from 'react';
 import { SignUp } from 'ui/signUp/SignUp';
 
@@ -26,7 +27,11 @@ jest.mock('ui/contexts', () => {
           displayConfig: {
             homeUrl: 'https://www.bbc.com',
           },
-          authConfig: { singleSessionMode: true } as Partial<AuthConfig>,
+          authConfig: {
+            identificationStrategies: ['email_address', 'oauth_google'],
+            firstFactors: ['email_address', 'oauth_google'],
+            singleSessionMode: true,
+          } as Partial<AuthConfig>,
         } as Partial<EnvironmentResource>),
     ),
     withCoreSessionSwitchGuard: (a: any) => a,