+++ packages/clerk-js/src/ui/signUp/SignUpStart.tsx
@@ -3,7 +3,12 @@ import { Form } from '@clerk/shared/components/form';
 import { Input } from '@clerk/shared/components/input';
 import { PhoneInput } from '@clerk/shared/components/phoneInput';
 import { noop } from '@clerk/shared/utils';
-import { SignUpParams, SignUpResource } from '@clerk/types';
+import {
+  OAuthStrategy,
+  SignUpParams,
+  SignUpResource,
+  Web3Strategy,
+} from '@clerk/types';
 import React from 'react';
 import type { FieldState } from 'ui/common';
 import {
@@ -30,14 +35,17 @@ import { getClerkQueryParam } from 'utils/getClerkQueryParam';
 import { SignInLink } from './SignInLink';
 import { SignUpOAuth } from './SignUpOAuth';
 import { SignUpWeb3 } from './SignUpWeb3';
-import { determineFirstPartyFields } from './utils';
+import {
+  determineFirstPartyFields,
+  determineOauthOptions,
+  determineWeb3Options,
+} from './utils';
 
 type ActiveIdentifier = 'emailAddress' | 'phoneNumber';
 
 function _SignUpStart(): JSX.Element {
   const { navigate } = useNavigate();
   const environment = useEnvironment();
-  const { userSettings } = environment;
   const { setSession } = useCoreClerk();
   const { navigateAfterSignUp } = useSignUpContext();
   const [emailOrPhoneActive, setEmailOrPhoneActive] =
@@ -73,8 +81,8 @@ function _SignUpStart(): JSX.Element {
     hasInvitationToken,
     hasOrganizationInvitationToken,
   );
-  const oauthOptions = userSettings.socialProviderStrategies;
-  const web3Options = userSettings.web3FirstFactors;
+  const oauthOptions = determineOauthOptions(environment) as OAuthStrategy[];
+  const web3Options = determineWeb3Options(environment) as Web3Strategy[];
 
   const handleTokenFlow = () => {
     const invitationToken = formFields.invitationToken.value;
@@ -94,6 +102,7 @@ function _SignUpStart(): JSX.Element {
         void completeSignUpFlow(res);
       })
       .catch(err => {
+        formFields.invitationToken.setValue('');
         handleError(err, [], setError);
       })
       .finally(() => {
@@ -179,7 +188,7 @@ function _SignUpStart(): JSX.Element {
     }
   };
 
-  const completeSignUpFlow = (su: SignUpResource) => {
+  const completeSignUpFlow = async (su: SignUpResource) => {
     if (su.status === 'complete') {
       return setSession(su.createdSessionId, navigateAfterSignUp);
     } else if (