+++ packages/clerk-js/src/ui/signUp/utils.ts
@@ -1,5 +1,9 @@
-import { snakeToCamel } from '@clerk/shared/utils';
-import type { Attributes, EnvironmentResource } from '@clerk/types';
+import type {
+  EnvironmentResource,
+  IdentificationStrategy,
+  SignInStrategyName,
+  ToggleTypeWithRequire,
+} from '@clerk/types';
 
 type FieldKeys =
   | 'emailOrPhone'
@@ -12,52 +16,69 @@ type FieldKeys =
   | 'invitationToken'
   | 'organizationInvitationToken';
 
-// TODO: Refactor SignUp component and remove
-// this leftover type
 type Fields = {
-  [key in FieldKeys]?: 'on' | 'off' | 'required';
+  [key in FieldKeys]?: ToggleTypeWithRequire;
 };
-
-function isEmailOrPhone(attributes: Attributes) {
-  return (
-    attributes.email_address.used_for_first_factor &&
-    attributes.phone_number.used_for_first_factor
-  );
-}
 
 export function determineFirstPartyFields(
   environment: EnvironmentResource,
   hasInvitation?: boolean,
   hasOrganizationInvitation?: boolean,
 ): Fields {
-  const { attributes } = environment.userSettings;
+  const idRequirements =
+    environment.authConfig.identificationRequirements.flat();
   const fields: Fields = {};
 
-  Object.entries(attributes)
-    .filter(([key]) => ['username', 'first_name', 'last_name'].includes(key))
-    .filter(([, desc]) => desc.enabled)
-    .forEach(
-      ([key, desc]) =>
-        (fields[snakeToCamel(key) as keyof Fields] = desc.required
-          ? 'required'
-          : 'on'),
-    );
-
+  const idByEmail = idRequirements.includes('email_address');
+  const idByPhone = idRequirements.includes('phone_number');
+  const idByEmailOrPhone = idByEmail && idByPhone;
+  const idByUsername =
+    idRequirements.includes('username') ||
+    environment.authConfig.username === 'on';
   if (hasInvitation) {
     fields.invitationToken = 'required';
   } else if (hasOrganizationInvitation) {
     fields.organizationInvitationToken = 'required';
-  } else if (isEmailOrPhone(attributes)) {
+  } else if (idByEmailOrPhone) {
     fields.emailOrPhone = 'required';
-  } else if (attributes.email_address.used_for_first_factor) {
+  } else if (idByEmail) {
     fields.emailAddress = 'required';
-  } else if (attributes.phone_number.used_for_first_factor) {
+  } else if (idByPhone) {
     fields.phoneNumber = 'required';
   }
 
-  if (attributes.password.required) {
-    fields.password = 'required';
+  if (idByEmailOrPhone || idByEmail || idByPhone) {
+    if (idByUsername) {
+      fields.username = environment.authConfig.username;
+    }
+
+    if (environment.authConfig.password === 'required') {
+      fields.password = environment.authConfig.password;
+    }
+
+    if (['on', 'required'].includes(environment.authConfig.firstName)) {
+      fields.firstName = environment.authConfig.firstName;
+    }
+    if (['on', 'required'].includes(environment.authConfig.lastName)) {
+      fields.lastName = environment.authConfig.lastName;
+    }
   }
 
   return fields;
+}
+
+export function determineOauthOptions(
+  environment: EnvironmentResource,
+): IdentificationStrategy[] {
+  const idRequirements = [
+    ...new Set(environment.authConfig.identificationRequirements.flat()),
+  ];
+  return idRequirements.filter(fac => fac.startsWith('oauth')).sort();
+}
+
+export function determineWeb3Options(
+  environment: EnvironmentResource,
+): SignInStrategyName[] {
+  const idRequirements = [...new Set(environment.authConfig.firstFactors)];
+  return idRequirements.filter(fac => fac.startsWith('web3')).sort();
 }