+++ packages/clerk-js/src/ui/userProfile/account/personalInformation/PersonalInformationCard.tsx
@@ -1,23 +1,12 @@
 import { List } from '@clerk/shared/components/list';
 import { TitledCard } from '@clerk/shared/components/titledCard';
 import React from 'react';
-import { useCoreUser, useEnvironment } from 'ui/contexts';
+import { useCoreUser } from 'ui/contexts';
 import { useNavigate } from 'ui/hooks';
 
-export const PersonalInformationCard = (): JSX.Element | null => {
+export const PersonalInformationCard = (): JSX.Element => {
   const user = useCoreUser();
   const { navigate } = useNavigate();
-  const { userSettings } = useEnvironment();
-  const {
-    attributes: { first_name, last_name },
-  } = userSettings;
-
-  const hasAtLeastOneAttributeEnable =
-    first_name?.enabled || last_name?.enabled;
-
-  if (!hasAtLeastOneAttributeEnable) {
-    return null;
-  }
 
   const firstNameRow = (
     <List.Item
@@ -41,9 +30,6 @@ export const PersonalInformationCard = (): JSX.Element | null => {
     </List.Item>
   );
 
-  const showFirstName = first_name.enabled;
-  const showLastName = last_name.enabled;
-
   return (
     <TitledCard
       className='cl-themed-card'
@@ -51,8 +37,8 @@ export const PersonalInformationCard = (): JSX.Element | null => {
       subtitle='Manage personal information settings'
     >
       <List className='cl-titled-card-list'>
-        {showFirstName && firstNameRow}
-        {showLastName && lastnameRow}
+        {firstNameRow}
+        {lastnameRow}
       </List>
     </TitledCard>
   );