+++ packages/clerk-js/src/ui/userProfile/emailAdressess/EmailDetail.test.tsx
@@ -1,14 +1,14 @@
 import { render, renderJSON, screen, userEvent } from '@clerk/shared/testUtils';
-import { EmailAddressResource, EnvironmentResource, SignInStrategyName, UserSettingsResource } from '@clerk/types';
+import { EmailAddressResource } from '@clerk/types';
+import { SignInStrategyName } from '@clerk/types';
 import { ClerkAPIResponseError } from 'core/resources/Error';
 import React from 'react';
-import { PartialDeep } from 'type-fest';
-import { EmailDetail } from './EmailDetail';
 
+import { EmailDetail } from './EmailDetail';
 
 const mockNavigate = jest.fn();
 const mockUseCoreUser = jest.fn();
-let mockFirstFactors = {} as PartialDeep<UserSettingsResource>;
+let mockFirstFactors = [] as SignInStrategyName[];
 
 jest.mock('ui/hooks', () => ({
   useNavigate: () => {
@@ -39,8 +39,11 @@ jest.mock('ui/contexts', () => {
   return {
     useCoreUser: () => mockUseCoreUser(),
     useEnvironment: jest.fn(() => ({
-      userSettings: mockFirstFactors
-    }) as PartialDeep<EnvironmentResource>),
+      authConfig: {
+        firstFactors: mockFirstFactors,
+        emailAddressVerificationStrategies: mockFirstFactors,
+      },
+    })),
     useUserProfileContext: jest.fn(() => {
       return {
         routing: 'path',
@@ -72,14 +75,7 @@ describe('<EmailDetail/>', () => {
   });
 
   it('displays verification errors', async () => {
-    mockFirstFactors = {
-      attributes: {
-        email_address: {
-          used_for_first_factor: true,
-          first_factors: ['email_code']
-        }
-      }
-    }
+    mockFirstFactors = ['email_code'];
     mockUseCoreUser.mockImplementation(() => {
       return {
         primaryEmailAddressId: '1',