+++ packages/clerk-js/src/ui/userProfile/security/index.tsx
@@ -19,11 +19,10 @@ export function SecurityRoutes({
   standAlone = false,
   index = false,
 }: SecurityRoutesProps): JSX.Element {
-  const { userSettings } = useEnvironment();
-  const { attributes: { phone_number } } = userSettings;
+  const { authConfig } = useEnvironment();
   const canActivate2svRoutes = () =>
-    phone_number.used_for_second_factor &&
-    phone_number.second_factors.includes('phone_code');
+    authConfig.secondFactors.length > 0 &&
+    authConfig.secondFactors.includes('phone_code');
 
   return (
     <Route