+++ packages/types/src/index.ts
@@ -22,7 +22,6 @@ export * from './signUp';
 export * from './theme';
 export * from './token';
 export * from './user';
-export * from './userSettings';
 export * from './utils';
 export * from './verification';
 export * from './web3';