+++ packages/types/src/json.ts
@@ -3,13 +3,17 @@
  */
 
 import { MembershipRole, OrganizationInvitationStatus } from '.';
+import { ToggleType, ToggleTypeWithRequire } from './authConfig';
+import { EmailAddressVerificationStrategy } from './emailAddress';
 import { OAuthStrategy } from './oauth';
 import { SessionStatus } from './session';
 import {
+  IdentificationStrategy,
   PreferredSignInStrategy,
   SignInFactor,
   SignInIdentifier,
   SignInStatus,
+  SignInStrategyName,
   UserData,
 } from './signIn';
 import { SignUpField, SignUpIdentificationField, SignUpStatus } from './signUp';
@@ -21,7 +25,6 @@ import {
   FontWeight,
   HexColor,
 } from './theme';
-import { UserSettingsJSON } from './userSettings';
 import { VerificationStatus } from './verification';
 
 export interface ClerkResourceJSON {
@@ -84,7 +87,6 @@ export interface ImageJSON {
 export interface EnvironmentJSON extends ClerkResourceJSON {
   auth_config: AuthConfigJSON;
   display_config: DisplayConfigJSON;
-  user_settings: UserSettingsJSON;
 }
 
 export interface ClientJSON extends ClerkResourceJSON {
@@ -214,9 +216,24 @@ export interface PublicUserDataJSON extends ClerkResourceJSON {
 export interface SessionWithActivitiesJSON extends Omit<SessionJSON, 'user'> {
   user: null;
   latest_activity: SessionActivityJSON;
+  // activities: SessionActivityJSON[];
 }
 
-export interface AuthConfigJSON extends ClerkResourceJSON {
+export interface AuthConfigJSON {
+  object: 'auth_config';
+  id: string;
+  first_name: ToggleTypeWithRequire;
+  last_name: ToggleTypeWithRequire;
+  email_address: ToggleType;
+  phone_number: ToggleType;
+  username: ToggleType;
+  password: ToggleTypeWithRequire;
+  identification_strategies: IdentificationStrategy[];
+  identification_requirements: IdentificationStrategy[][];
+  password_conditions: any;
+  first_factors: SignInStrategyName[];
+  second_factors: SignInStrategyName[];
+  email_address_verification_strategies: EmailAddressVerificationStrategy[];
   single_session_mode: boolean;
 }