+++ packages/expo/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.8.7](https://github.com/clerkinc/javascript/compare/@clerk/clerk-expo@0.8.6...@clerk/clerk-expo@0.8.7) (2022-03-01)
+
+**Note:** Version bump only for package @clerk/clerk-expo
+
+
+
+
+
+### [0.8.6](https://github.com/clerkinc/javascript/compare/@clerk/clerk-expo@0.8.6-staging.0...@clerk/clerk-expo@0.8.6) (2022-02-24)
+
+**Note:** Version bump only for package @clerk/clerk-expo
+
+
+
+
+
+### [0.8.6-staging.0](https://github.com/clerkinc/javascript/compare/@clerk/clerk-expo@0.8.5-staging.0...@clerk/clerk-expo@0.8.6-staging.0) (2022-02-24)
+
+**Note:** Version bump only for package @clerk/clerk-expo
+
+
+
+
+
+### [0.8.5-staging.0](https://github.com/clerkinc/javascript/compare/@clerk/clerk-expo@0.8.4-staging.0...@clerk/clerk-expo@0.8.5-staging.0) (2022-02-22)
+
+**Note:** Version bump only for package @clerk/clerk-expo
+
+
+
+
+
 ### [0.8.4-staging.0](https://github.com/clerkinc/javascript/compare/@clerk/clerk-expo@0.8.3-staging.1...@clerk/clerk-expo@0.8.4-staging.0) (2022-02-17)
 
 **Note:** Version bump only for package @clerk/clerk-expo