+++ packages/types/CHANGELOG.md
@@ -3,6 +3,37 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.26.0](https://github.com/clerkinc/javascript/compare/@clerk/types@1.25.4...@clerk/types@1.26.0) (2022-03-01)
+
+
+### Features
+
+* **types:** Add support for oauth_microsoft ([96c1cc6](https://github.com/clerkinc/javascript/commit/96c1cc6817b9bbc6917ea2773498299c1ff9b951))
+
+
+
+### [1.25.4](https://github.com/clerkinc/javascript/compare/@clerk/types@1.25.4-staging.0...@clerk/types@1.25.4) (2022-02-24)
+
+**Note:** Version bump only for package @clerk/types
+
+
+
+
+
+### [1.25.4-staging.0](https://github.com/clerkinc/javascript/compare/@clerk/types@1.25.3-staging.0...@clerk/types@1.25.4-staging.0) (2022-02-24)
+
+
+### Features
+
+* **clerk-js:** Introduce `UserSettings.instanceIsPasswordBased` ([f72a555](https://github.com/clerkinc/javascript/commit/f72a555f6adb38870539e9bab63cb638c04517d6))
+
+
+### Bug Fixes
+
+* **clerk-js,clerk-react:** Revert user settings work ([9a70576](https://github.com/clerkinc/javascript/commit/9a70576d1a47f01e6dbbfd8704f321daddcfe590))
+
+
+
 ### [1.25.3-staging.0](https://github.com/clerkinc/javascript/compare/@clerk/types@1.25.2-staging.0...@clerk/types@1.25.3-staging.0) (2022-02-17)
 
 **Note:** Version bump only for package @clerk/types